requirements.txt
openai
python-dotenv
requests
pyaudio
numpy
fastapi
uvicorn
jinja2
python-multipart
aiofiles
sqlalchemy
databases[sqlite]
aiosqlite